Save changes

+++ patch_review_start.sh
@@ -1,25 +1,32 @@
-#!/bin/bash -ex
+#!/bin/bash
+set -ex
 
 curr_branch=$(git branch | grep "^* " | awk '{print $2}')
-echo "${curr_branch}" > .orig_branch
 
-base_branch="master"
-if [ "X${curr_branch}" != "X" ]; then
-	if (git branch | grep -qw "master.${curr_branch}"); then
-		base_branch="master.${curr_branch}"
+if [ "X${curr_branch}" == "Xreviewbranch" ]; then
+	echo "-I- You are already on review branch, will only run patchreview tool"
+else
+	echo "-I- Preparing stuff before running patchreview tool..."
+
+	echo "${curr_branch}" > .orig_branch
+
+	base_branch="master"
+	if [ "X${curr_branch}" != "X" ]; then
+		if (git branch | grep -qw "master.${curr_branch}"); then
+			base_branch="master.${curr_branch}"
+		fi
 	fi
-fi
 
-/bin/rm -rf toreview
-git format-patch ${base_branch} -o toreview
-git co ${base_branch} -B reviewbranch
+	/bin/rm -rf toreview
+	git format-patch ${base_branch} -o toreview
+	git co ${base_branch} -B reviewbranch
 
-/bin/rm -rf toreview_work
-mkdir toreview_work
+	/bin/rm -rf toreview_work
+	mkdir toreview_work
 
-/bin/rm -rf .data
-mkdir .data
-cat > .data/patchreview.prj << EOF
+	/bin/rm -rf .data
+	mkdir .data
+	cat > .data/patchreview.prj << EOF
 0
 1
 1
@@ -30,5 +37,6 @@ true
 true
 true
 EOF
+fi
 
 patchreview
